remove plugins dir
