+++ src/components/ChannelsAvailabilityMenuContent/ChannelsAvailabilityMenuContent.tsx
@@ -1,13 +1,13 @@
 // @ts-strict-ignore
 import HorizontalSpacer from "@dashboard/components/HorizontalSpacer";
 import { CollectionFragment } from "@dashboard/graphql";
-import { ScrollableContent } from "@dashboard/plugins/components/PluginsList/PluginAvailabilityStatusPopup/ScrollableContent";
 import { PillColor } from "@saleor/macaw-ui";
 import { Text } from "@saleor/macaw-ui-next";
 import { MessageDescriptor, useIntl } from "react-intl";
 
 import { messages } from "../ChannelsAvailabilityDropdown/messages";
 import { Pill } from "../Pill";
+import { ScrollableContent } from "./PluginAvailabilityStatusPopup/ScrollableContent";
 import { useStyles } from "./styles";
 
 interface ChannelsAvailabilityMenuContentProps {
