+++ src/components/ChannelsAvailabilityMenuContent/PluginAvailabilityStatusPopup/ScrollableContent.tsx
