+++ src/extensions/mutations.ts
@@ -51,3 +51,16 @@ export const webhookDelete = gql`
     }
   }
 `;
+
+export const pluginUpdate = gql`
+  mutation PluginUpdate($channelId: ID, $id: ID!, $input: PluginUpdateInput!) {
+    pluginUpdate(channelId: $channelId, id: $id, input: $input) {
+      errors {
+        ...PluginError
+      }
+      plugin {
+        ...PluginsDetails
+      }
+    }
+  }
+`;
