--- src/plugins/urls.ts
@@ -1,4 +1,0 @@
-import { Dialog, SingleAction } from "../types";
-
-type PluginUrlDialog = "clear" | "edit";
-export type PluginUrlQueryParams = Dialog<PluginUrlDialog> & SingleAction;
